+++ src/main/java/org/betonquest/betonquest/events/ExperienceEvent.java
@@ -3,10 +3,10 @@ package org.betonquest.betonquest.events;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.VariableNumber;
 import org.betonquest.betonquest.api.QuestEvent;
+import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.exceptions.QuestRuntimeException;
-import org.bukkit.entity.Player;
 
 /**
  * Gives the player specified amount of experience
@@ -14,6 +14,10 @@ import org.bukkit.entity.Player;
 @SuppressWarnings("PMD.CommentRequired")
 public class ExperienceEvent extends QuestEvent {
 
+    /**
+     * The experience level the player needs to get.
+     * The decimal part of the number is a percentage of the next level.
+     */
     private final VariableNumber amount;
     private final boolean checkForLevel;
 
@@ -25,13 +29,20 @@ public class ExperienceEvent extends QuestEvent {
 
     @Override
     protected Void execute(final Profile profile) throws QuestRuntimeException {
-        final Player player = profile.getOnlineProfile().get().getPlayer();
-        final int amount = this.amount.getInt(profile);
-        if (checkForLevel) {
-            player.giveExpLevels(amount);
-        } else {
-            player.giveExp(amount);
-        }
+        final double amount = this.amount.getDouble(profile);
+
+        profile.getOnlineProfile()
+                .map(OnlineProfile::getPlayer)
+                .ifPresent(player -> {
+                    if (checkForLevel) {
+                        final double current = player.getLevel() + player.getExp();
+                        final double amountToAdd = current + amount;
+                        player.setLevel((int) amountToAdd);
+                        player.setExp((float) (amountToAdd - (int) amountToAdd));
+                    } else {
+                        player.giveExp((int) amount);
+                    }
+                });
         return null;
     }
 }